python/pyproject.toml
[project]
name = "sigil-atlas"
version = "0.1.0"
description = "SigilAtlas Python sidecar — ingest pipeline and embedding"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "open-clip-torch>=2.24",
    "Pillow>=10.0",
    "exifread>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
